fix(cli): Allow using current version as baseline
In most cases, people do not bump the version post-release so the
current version in the manifest is actually the baseline.

This will cause problems when we add support for running against
dependencies but I think that can be resolved by only doing `<=` for
workspace members and `<` for everything else.

Fixes #118