Optimize the `#[must_use]` lint queries by improving filtering.
Most items are not `#[must_use]`. By moving the more-selective branch
of the query (the `current` one that looks for `#[must_use]`),
the query will evaluate fewer items because it will discard
more of them earlier in the process.

This is an optimization that requires knowing the underlying shape
of the dataset being queried, so it's not something that Trustfall
could have applied automatically without access to statistics.