Optimize the `#[must_use]` lint queries by improving filtering. (#286)
Most items are not `#[must_use]`. By moving the more-selective branch of the query (the `current` one that looks for `#[must_use]`), the query will evaluate fewer items because it will discard more of them earlier in the process.

The speedup is approximately a factor of 1600-2000x when measured on a particularly large crate: `aws-sdk-ec2`
https://github.com/awslabs/aws-sdk-rust/tree/main/sdk/ec2 

Running just the 4 affected lints on that crate:
- Before: 401-413s total scan time
- After: 0.21-0.25s total scan time

The numbers only include lint execution time, excluding things such as the time needed to build the rustdoc JSON.

This is an optimization that requires knowing the underlying shape of the dataset being queried, so it's not something that Trustfall could have applied automatically without access to statistics.