Remove ability to check bin-only targets, to fix `doc = false` issue.
Because of a complex web of edge cases, we can _either_ choose to continue to support our legacy behavior of allowing bin-only targets to be semver-checked, _or_ we can support semver-checking lib targets that set `doc = false`. Semver-checking bin-only targets doesn't make a lot of sense, since they cannot be used as a library anyway. This functionality was always slated for removal, I just wish we had more time to deprecate & warn on it before just ripping it out like this.

Now we're in a tight spot, and this is definitely the lesser evil -- we're making the tool work for library users who produce an API that can reasonably be checked and used as such.

Fixes #660.