Explicitly `cargo update` in path dependency workspaces. (#901)
Resolves a variant of #167 described here:
https://github.com/obi1kenobi/cargo-semver-checks/issues/167#issuecomment-2324959305

This issue was observed during a CI run of #900, where an older
tokio-stream version was using a brand-new tokio with stronger
guarantees: several types had newly become `UnwindSafe`.

Since the baseline tokio-stream was a registry version, it was using
a fresh lockfile with the latest dependencies. The current arm,
meanwhile, was a path dependency fixed to a specific commit, where
cargo had decided to reuse an older version of tokio without the
`UnwindSafe` trait on those types.

This presented as a (phantom) breaking change: the older tokio-stream's
types were `UnwindSafe` due to auto-trait propagation from a newer tokio
while the newer tokio-stream with an older tokio did not have that trait.

The solution is to explicitly run `cargo update` inside path dependency
workspaces we create. This way, both path dependency and index-based
rustdoc JSON generation happens with the latest versions of the target
library's own dependencies.