Support linting breaking changes in manifests & add `feature_missing` lint (#1007)
Add support for linting of package manifests, allowing us to scan for
breaking changes there as well.

For example, deleting a feature is a major breaking change. As of this
PR, we can detect and report that:
```
--- failure feature_missing: package feature removed or renamed ---

Description:
A feature has been removed from this package's Cargo.toml. This will break downstream crates which enable that feature.
        ref: https://doc.rust-lang.org/cargo/reference/semver.html#cargo-feature-remove
       impl: https://github.com/obi1kenobi/cargo-semver-checks/tree/v0.37.0/src/lints/feature_missing.ron

Failed in:
  feature going_missing in the package's Cargo.toml
  feature rand in the package's Cargo.toml

     Summary semver requires new major version: 1 major and 0 minor checks failed
```

Completes the first checkbox of the 2024H2 Rust Project Goal on
cargo-semver-checks:
https://github.com/rust-lang/rust-project-goals/issues/104

Unblocks the lints specified in #48.