Add two lints looking for new fields in `repr(C)` unions. (#1049)
If a union is `repr(C)`, downstream users may rely on bit-validity
invariants that may or may not be intended by the union type. As
described in #950, we differentiate between two cases here based on
whether the bit-validity invariant is reasonably implied by the public
API, or whether it might be the product of Hyrum's Law.

This PR also discovered and fixes a small bug in the
`union_pub_field_now_doc_hidden` lint, which could have a false-positive
on fields that are no longer `pub` but never became `#[doc(hidden)]`.

Resolves #950.