Require test crates to pass  `cargo check`. (#1145)
It seems not all our test crates pass `cargo check` since rustdoc (with
good reasons) allows certain compile errors to go unobserved.

In a couple of our test crates, that's expected and desirable. For
example, we have test cases for what happens when we lint a crate that
doesn't build.

In most cases, that's a bug in the test that we should fix.