feat: add a witness hint for `function_parameter_count_changed` lint (#1194)
I’ve tried to keep the hint simple and consistent with
others(#978,#977):
`parameter_count_changed::function_with_parameters_removed(#parameter !=
<old_count>);`

However, the syntax `(#parameter != <old_count>)` or as alternative
`(|parameter| != <old_count>)` might be confusing when compared to
standard Rust syntax. Would it be better to replace this with `(...)` or
`(...,...,...)` where number of `...` be equal to old_parameter_count
for clarity?

>I'm still getting used to reading and writing these, so please forgive
any mistakes.

---------

Co-authored-by: Predrag Gruevski <2348618+obi1kenobi@users.noreply.github.com>