Chore: Add witness hint for `enum_struct_variant_field_marked_deprecated` lint (#1209)
Adds a witness hint for the
`enum_struct_variant_field_marked_deprecated` lint.

## Example
`enum_struct_variant_field_marked_deprecated` outputs the following
witness hint for the
`./test_crates/enum_struct_variant_field_marked_deprecated/` test.
```rust
#[deny(deprecated)]
match value {
    enum_struct_variant_field_marked_deprecated::NormalEnum::StructVariant { field_to_be_deprecated, .. } => (),
    _ => (),
}
```
This compiles on the old version, but not on the new due to the presence
of `#[deny(deprecated)]`, as when`field_to_be_deprecated` becomes marked
as `deprecated`, compilation fails.

## Comments
I hope using something like `#[deny(deprecated)]` is ok for witnesses,
as that was the only practical way I could think of to display the issue
in this case.