Chore: Add witness hint for `enum_struct_variant_field_added` lint (#1208)
Adds a witness hint for the `enum_struct_variant_field_added` lint.

## Example
`enum_struct_variant_field_added` outputs the following witness hint for
the `./test_crates/enum_struct_variant_field_added/` test.
```rust
match value {
    enum_struct_variant_field_added::PubEnum::Foo { x } => (),
    _ => (),
}
```
This compiles on the old version, but not on the new due to the lack of
exhaustiveness in the destructuring of the enum variant.

## Commit Notes
+ Added witness hint for `enum_struct_variant_field_added`
+ Added new parts to query for `enum_struct_variant_field_added` to
fetch relevant data for the witness

## Comments
I had to add more to the `trustfall` query for
`enum_struct_variant_field_added`, which took a little bit to figure
out, but I'm starting to get the hang of it. Is there anywhere that
would be good to read up on the syntax and whatnot for `trustfall`? I've
found a few myself, but more resources are always welcome.

Aditionally, feel free to give me any feedback on this one, especially
surrounding the new parts of the query. This is what I found that
worked, but if there's a better, or more idiomatic way to do this, I
would appreciate knowing!