Chore: Add witness hint for `enum_tuple_variant_changed_kind` lint (#1211)
Adds a witness hint for the `enum_tuple_variant_changed_kind` lint.

## Example
`enum_tuple_variant_changed_kind` outputs the following witness hint for
one of the `./test_crates/enum_tuple_variant_changed_kind/` tests.
```rust
match value {
    enum_tuple_variant_changed_kind::MultipleTest::WillBecomeStructLike(..) => (),
    _ => (),
}
```
This compiles on the old version, but not on the new due to the presence
of tuple destructuring, as when it changes type (to struct-like or
unit-like), it can no longer be destructured as a tuple.