Feat: Add custom `repeat` handlebars helper (#1219)
Adds a custom handlebars block helper called `repeat`. The goal of this
helper is to allow one to iterate a value `n` number of times. The
syntax of `repeat` is similar to that of the built-in `each` helper. It
assigns to the local block variables `@index`, `@first`, and `@last`.
`@index` is a number, `@first` and `@last` are booleans.

## Example
The following example handlebars template assumes the existence of an
outputted value called `field_count`, which is a `Number` type
representing the number of fields on an enum variant (for example).
```
"path_to_enum::ExampleEnum::ExampleVariant({{#repeat field_count }}_{{#unless @last }}, {{/unless}}{{/times}})"
```
This template string would produce the following output string for an
input tuple enum with 3 fields:
```
"path_to_enum::ExampleEnum::ExampleVariant(_, _, _)"
```

## Commit Notes
+ Adds a #repeat block helper, similar to the built-in #each helper

## Comments
This PR is to support the progress of #1212, #1214, and #1216. I'm
creating this because as far as I am aware, there is not currently a way
to iterate based on a Number in the existing set of handlebars helpers.
Let me know if there's any improvements I can make to this!

---------

Co-authored-by: Predrag Gruevski <2348618+obi1kenobi@users.noreply.github.com>