Feat: Add custom `to_string` handlebars helper (#1236)
Adds a `to_string` custom handlebars helper.

## Comments
There's not much to show for this one, it just converts a Number or Bool
to a String Value. This isn't useful most of the time, as templating
converts a variable to a string anyways. It's main use case is for
comparing a number and a string. This is currently needed in #1214 and
#1216, as although we can now iterate using the `repeat` helper, it
produces Number Values, while `field_name`, which `@index` is compared
against, is a String Value, which will always fail the `eq` test. This
helper exists to resolve that issue.

---------

Co-authored-by: Predrag Gruevski <2348618+obi1kenobi@users.noreply.github.com>