[Witness Generation] Add unstable witness program flag (#1276)
# Commit Notes
+ Added new unstable flag `witnesses`
+ Flag requires output path `PathBuf` for target witness output

# Comments
Feels good to get started! This just adds an unstable flag as was
suggested on the rust-lang Zulip to gate the witness generation system.
Decided to make the flag require an output directory, since that is
mandatory for the witness system to work anyways.

Nothing is done with the flag as of right now, it just exists.