[Witness Generation]  Add `InheritedValue` inheritance (#1330)
# Commit Notes
+ Adds the `InheritOutput` variant to `InheritedValue`
+ Adds a function to inherit prior values for witness generation
# Comments
Technically this does the inheritance for both arguments and outputs,
which is needed for witness querying systems. I'm not happy with the
cloning I put, that was just to get something in and working. Also, I'd
prefer to use an `Iterator` instead of a for loop (allows use of
`BTreeMap::from_iter` which would be nice), just makes it nicer to read
in my mind. Generally just want some feedback on this, but overall I'm
pretty happy with this as at least a first iteration of the value
inheritance.

Going to put this as a draft PR at least for now, since it doesn't
really *need* to be merged right away and I'd like to improve it. This
is just getting something hammered out. If you want to merge this
anyways, just so we have something, that also works.

Also, not entirely happy with all my variable names, function names,
etc. feel free to modify/suggest away.