[Witness Generation] Add witness query system & witness text generation (#1342)
# Commit Notes
+ Add new `witness_gen` module
+ Add logic to run witness queries
+ Witness query data is unused as of right now

# Comments
This took longer than I thought it would. Not the programming itself but
life has not been on the side of my time lately.

Either way, this adds a new `witness_gen` module which I plan to use for
a lot more of the witness generation system. Here, I just wanted it to
separate things out a little bit. Currently, I'm not doing anything with
the query result, and in fact, I haven't had the opportunity to test if
it even does *anything* but I'll try to quickly test that next, I'd like
to know it's working before going on to generating the actual witnesses
from this data. Also, witness queries that fail for some reason or
another currently keep their own errors, I'd rather design the system to
allow as many witnesses to be generated as possible, and then simply
report error/gather diagnostics for any query or witness that fails.
Unlike the lint queries themselves, it arguably should not fail hard,
and should allow the rest of `c-s-c` to complete, it should just report
any errors that occur to the user, and prompt them to report it as an
issue. At least, that's my opinion on it, thoughts?

Additionally, as I state in one of the comments, if no witness query
exists, we ideally want to forward existing lint_results, so those
values can still be used in a witness that may not need a query. Though
looking at this now, it might be optimal to add another case to the
`match` for specifically where there *is* a witness to generate and
*not* a witness query present to do that, and then on the catchall,
instead just get filtered out (switching to filter_map instead of just
map would allow this). However, this does still work, it's just a
possible improvement for later.

---------

Co-authored-by: Predrag Gruevski <2348618+obi1kenobi@users.noreply.github.com>