Refactor rustdoc generation, replacing dynamic dispatch system (#1354)
# Commit Notes

+ Adds new `generate_data_request` method on `RustdocGenerator` trait
+ Changes `load_rustdoc` method to require optional `CrateDataRequest`
+ Splits rustdoc loading into two steps, which raises up the
`CrateDataRequest`, making it accessible elsewhere

# Comments

This almost definitely needs more work, as this was just a crude hashing
of what needs to happen. I'm going to put this as a draft for now, just
to get feedback, but it's far from perfect in my opinion.

What this *does* do is breaks out the `CrateDataRequest` type from being
a deeply internal type to being something accessible at the top level,
allowing for easy access to information about the crate being queried.
Now, the value returned is optional, since feeding rustdoc directly does
not create a `CrateDataRequest`, that's why there's
`Option<CrateDataRequest<'_>>` in quite a few places.

This is mainly a necessary dependency for witness generation, as the
generation of a witness crate requires knowing information about where
the crate exists, or that it exists on crates.io.