Make the -q command-line option useful.
As of version 0.45.0, using the -q option on the
command-line nearly completely eliminates all console
output, which makes it not particularly useful.

This PR makes it so when using the -q option,
any lint violations are still printed to the console.
This is the mode I'd like to use in our CI system.